code refresh

+++ src/SQRT.java
@@ -17,12 +17,12 @@ public class SQRT {
         if( y < 0) throw new IllegalArgumentException("Cannot be negative");
         
         double e = 0.000000001f;
-        double x0 = y, x1 = 0.0;
+        float x0 = y, x1 = 0.0f;
         while( Math.abs( (x0 - x1) / x0 ) > e){
             x0 = (x0 + x1)/2;
             x1 = y / x0;
         }
-        return (float)x1;
+        return x1;
     }
 
 }
